+++ src/App.tsx
@@ -1,31 +1,14 @@
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-
-import {
-  authRoutes,
-  designListRoutes,
-  homeRoutes,
-  myPageRoutes,
-  storeRoutes,
-  viewRoutes,
-} from '@routes';
-
 import queryClient from './queryClient';
+import { pageRoutes } from '@routes';
 
 import '@styles/global.css';
-const allRoutes = [
-  ...authRoutes,
-  ...designListRoutes,
-  ...homeRoutes,
-  ...myPageRoutes,
-  ...storeRoutes,
-  ...viewRoutes,
-];
-
-const router = createBrowserRouter(allRoutes);
 
 const App = () => {
+  const router = createBrowserRouter(pageRoutes);
+
   return (
     <QueryClientProvider client={queryClient}>
       <RouterProvider router={router} />
