+++ src/routes/layout/Layout.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const layoutStyle = style({
+  paddingTop: '6rem',
+});
